# PREF: Speed up `to_set`