--- a/dtoolkit/accessor/series/to_set.py
+++ b/dtoolkit/accessor/series/to_set.py
@@ -36,4 +36,4 @@ def to_set(s: pd.Series) -> set:
     {1, 2}
     """
 
-    return set(s.unique())
+    return set(s.to_list())